+++ Lab1/Lab1/Canvas.h
@@ -3,17 +3,17 @@
 class Canvas
 {
 public:
-	Canvas(uint8_t height, uint8_t width);
+	Canvas(int height, int width);
 	void repaint();
-	bool moveObject(uint8_t id, uint8_t x, uint8_t y);
-	bool eraseObject(uint8_t id);
-	bool fillObject(uint8_t id);
+	bool moveObject(int id, int x, int y);
+	bool eraseObject(int id);
+	bool fillObject(int id);
 	void addObject(IFigure*);
 	inline std::vector<IFigure*> getObjects() const { return objects;};
 private:
 	std::vector<IFigure*> objects;
-	uint8_t width;
-	uint8_t height;
+	int width;
+	int height;
 };
 
 //237*60
